\name{static.mar.2007}
\alias{static.mar.2007}
\docType{data}
\title{Static Descriptive Data.}
\description{
Static descriptive data for stocks.
}
\usage{data(static.mar.2007)}
\format{
  A data frame with 4091 observations on the following 3 variables.
  \describe{
    \item{\code{id}}{a character vector}
    \item{\code{symbol}}{a character vector}
    \item{\code{name}}{a character vector}
  }
}
\examples{
data(static.mar.2007)
}
\keyword{datasets}
