\name{trade.mar.2007}
\alias{trade.mar.2007}
\docType{data}
\title{Trading data for March 2007}
\description{
 This is trading data for march 2007
}
\usage{data(trade.mar.2007)}
\format{
  A data frame with 442 observations on the following 5  variables.
  \describe{
    \item{\code{period}}{a Date}
    \item{\code{id}}{a character vector}
    \item{\code{side}}{a character vector}
    \item{\code{exec.qty}}{a numeric vector}
    \item{\code{exec.price}}{a numeric vector}
  }
}
\examples{
data(trade.mar.2007)
}
\keyword{datasets}
