\name{tradeCosts-package}
\alias{tradeCosts-package}
\docType{package}

\title{
  Analysing trade cost
}
\description{
  Classes for analysing trade cost.
}
\details{
  \tabular{ll}{
    Package: \tab tradeCosts\cr
    Type: \tab Package\cr
    Version: \tab 1.0\cr
    Date: \tab 2007-06-06\cr
    Depends: \tab R (>= 2.4.0), methods, xtable, tools\cr
    License: \tab GPL version 2 or later\cr
  }
}

\author{
  Aaron Schwartz, Luyi Zhao, Jeff Enos, and David Kane
  Maintainer: Jeff Enos <jeff@kanecap.com>
}
\keyword{package}
