\name{dynamic.mar.2007}
\alias{dynamic.mar.2007}
\docType{data}
\title{Dynamic Descriptive Data for March}
\description{
This data set contains dynamic descriptive data for the month of march on
stocks with 5 different variables.
}
\usage{data(dynamic.mar.2007)}
\format{
  A data frame with 1688 observations on the following 5  variables.
  \describe{
    \item{\code{period}}{a Date}
    \item{\code{id}}{a character vector}
    \item{\code{price}}{a numeric vector}
    \item{\code{vwap}}{a numeric vector}
    \item{\code{adjustment.factor}}{a numeric vector}
  }
}
\examples{
data(dynamic.march.2007)
}
\keyword{datasets}
