\name{static.mar.2007}
\alias{static.mar.2007}
\docType{data}
\title{Sample static descriptive data}
\description{
  Sample static descriptive data, or data that doesn't change over time,
  for a set of securities.
}
\usage{data(static.mar.2007)}
\format{
  A data frame with 4091 observations on the following 4 variables.
  \describe{
    \item{\code{id}}{a character vector specifying a unique security id}
    \item{\code{symbol}}{a character vector containing the security symbol}
    \item{\code{name}}{a character vector containing company name}
    \item{\code{sector}}{a character vector containing company sector}
  }
}
\examples{
data(static.mar.2007)
}
\keyword{datasets}
