\name{tradeCosts-package}
\alias{tradeCosts-package}
\alias{tradeCosts}
\docType{package}
\title{
Trade Cost Analysis
}
\description{
Performs ex-post analysis of security transaction costs.
}
\details{
\tabular{ll}{
Package: \tab tradeCosts\cr
Version: \tab 0.3-1\cr
Date: \tab 2010-02-18\cr
LazyLoad: \tab yes\cr
Depends: \tab R (>= 2.4.0), methods, tools, graphics\cr
License: \tab GPL (>= 2)\cr
}

Index:
\preformatted{
dynamic.mar.2007        Sample dynamic descriptive data
static.mar.2007         Sample static descriptive data
trade.costs             User-level function to perform basic trade cost
                        analysis.
trade.mar.2007          Sample trading data
tradeCosts-class        Class "tradeCosts"
tradeCosts-package      Trade Cost Analysis
tradeCostsResults-class
                        Class "tradeCostsResults"
}

Further information is available in the following vignettes:
\tabular{ll}{
\code{tradeCosts-article} \tab Trade Costs (source, pdf)\cr
\code{tradeCosts} \tab Using the tradeCosts package (source, pdf)\cr
}
}
\author{
Arjun Ravi Narayan <contact@arjunnarayan.com>, Aaron Schwartz
<Aaron.J.Schwartz@williams.edu>, Daniel Suo
<danielsuo@gmail.com> and Luyi Zhao <luyizhao@fas.harvard.edu>
with contributions from Jeff Enos <jeff@kanecap.com> and David
Kane <david@kanecap.com>

Maintainer: Jeff Enos <jeff@kanecap.com>
}
\keyword{ package }
